backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==3.1.0
Flask-Cors==5.0.0
numpy==2.0.1
opencv-python==4.10.0.84
pillow==10.4.0
ultralytics==8.3.31
gunicorn


# requirements.txt
python-dotenv
google-generativeai